refactor: rename Links to FgLink for clarity; move to widgets dir
