+++ src/components/ui/JobsPage/TicketRow.tsx
@@ -9,7 +9,7 @@ import {
   makeMapKey
 } from '@/utils';
 import { FileSharePath } from '@/shared.types';
-import { FgStyledLink } from '../Links';
+import { FgStyledLink } from '../widgets/FgLink';
 
 export default function TicketRow({ item }: { item: Ticket }) {
   const { zonesAndFileSharePathsMap } = useZoneAndFspMapContext();
